fix jshint
* Fix the gulpfile to actually test the source in lib.
* Refactored Control, floatingLayer, and fullscreen to be in a single
  module to resolve a hidden circular dependency issue.
* Fixed various jshint warnings

Issue: ENYO-2356
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)
