+++ lib/Control/Control.js
@@ -6,24 +6,26 @@ require('enyo');
 */
 
 var
-	kind = require('./kind'),
-	utils = require('./utils'),
-	platform = require('./platform'),
-	dispatcher = require('./dispatcher'),
-	fullscreen = require('./fullscreen'),
-	options = require('./options'),
-	roots = require('./roots');
+	kind = require('../kind'),
+	utils = require('../utils'),
+	platform = require('../platform'),
+	dispatcher = require('../dispatcher'),
+	options = require('../options'),
+	roots = require('../roots');
 
 var
-	AccessibilitySupport = require('./AccessibilitySupport'),
-	UiComponent = require('./UiComponent'),
-	HTMLStringDelegate = require('./HTMLStringDelegate'),
-	Dom = require('./dom'),
+	AccessibilitySupport = require('../AccessibilitySupport'),
+	UiComponent = require('../UiComponent'),
+	HTMLStringDelegate = require('../HTMLStringDelegate'),
+	Dom = require('../dom');
+
+var
+	fullscreen = require('./fullscreen'),
 	FloatingLayer = require('./floatingLayer');
 
 // While the namespace isn't needed here, gesture is required for ontap events for which Control
 // has a handler. Bringing them all in now for the time being.
-require('./gesture');
+require('../gesture');
 
 var nodePurgatory;
 
@@ -897,7 +899,7 @@ var Control = module.exports = kind(
 	* @public
 	*/
 	isFullscreen: function () {
-		return (this.hasNode() && this.node === fullscreen.getFullscreenElement());
+		return (this.hasNode() && this.node === Control.Fullscreen.getFullscreenElement());
 	},
 
 	/**
@@ -912,7 +914,7 @@ var Control = module.exports = kind(
 	requestFullscreen: function () {
 		if (!this.hasNode()) return false;
 
-		if (fullscreen.requestFullscreen(this)) {
+		if (Control.Fullscreen.requestFullscreen(this)) {
 			return true;
 		}
 
@@ -929,7 +931,7 @@ var Control = module.exports = kind(
 	*/
 	cancelFullscreen: function() {
 		if (this.isFullscreen()) {
-			fullscreen.cancelFullscreen();
+			Control.Fullscreen.cancelFullscreen();
 			return true;
 		}
 
@@ -1687,8 +1689,8 @@ Control.concat = function (ctor, props, instance) {
 
 Control.prototype.defaultKind = Control;
 
-// Control has to be *completely* set up before creating the floating layer otherwise things will
-// error out during construction.
+// Control has to be *completely* set up before creating the floating layer setting up the
+// fullscreen object because fullscreen depends on floating layer which depends on Control.
 
 /**
 * @static
@@ -1701,3 +1703,9 @@ Control.FloatingLayer = FloatingLayer(Control);
 * @public
 */
 Control.floatingLayer = new Control.FloatingLayer({id: 'floatingLayer'});
+
+/**
+* @static
+* @public
+*/
+Control.Fullscreen = fullscreen(Control);
