+++ lib/Control/floatingLayer.js
@@ -1,20 +1,18 @@
 /**
-* Exports the {@link module:enyo/floatingLayer~FloatingLayer} singleton instance.
-* @module enyo/floatingLayer
+* Exports the {@link module:enyo/Control/floatingLayer~FloatingLayer} singleton instance.
+* @module enyo/Control/floatingLayer
 */
 
-require('enyo');
-
 var
-	kind = require('./kind'),
-	platform = require('./platform');
+	kind = require('../kind'),
+	platform = require('../platform');
 
 module.exports = function (Control) {
 	/**
-	* {@link module:enyo/floatingLayer~FloatingLayer} is a [control]{@link module:enyo/Control~Control} that provides a layer for
-	* controls that should be displayed above an [application]{@link module:enyo/Application~Application}. The
-	* `floatingLayer` singleton can be set as a control's parent to have the control float
-	* above the application, e.g.:
+	* {@link module:enyo/Control/floatingLayer~FloatingLayer} is a
+	* [control]{@link module:enyo/Control~Control} that provides a layer for controls that should be
+	* displayed above an [application]{@link module:enyo/Application~Application}. The `floatingLayer`
+	* singleton can be set as a control's parent to have the control float above the application, e.g.:
 	*
 	* ```
 	* var floatingLayer = require('enyo/floatingLayer');
@@ -34,8 +32,8 @@ module.exports = function (Control) {
 	* @ui
 	* @protected
 	*/
-	return kind(
-		/** @lends module:enyo/floatingLayer~FloatingLayer.prototype */ {
+	var FloatingLayer = kind(
+		/** @lends module:enyo/Control/floatingLayer~FloatingLayer.prototype */ {
 
 		/**
 		* @private
@@ -116,4 +114,6 @@ module.exports = function (Control) {
 		teardownChildren: function () {
 		}
 	});
+
+	return FloatingLayer;
 };
