+++ lib/RichText/RichText.js
@@ -12,7 +12,9 @@ var
 	options = require('../options'),
 	utils = require('../utils'),
 	platform = require('../platform');
+
 var
+	Control = require('../Control'),
 	Input = require('../Input'),
 	RichTextAccessibilitySupport = require('./RichTextAccessibilitySupport');
