+++ lib/Scroller/Scroller.js
@@ -9,7 +9,6 @@ require('enyo');
 
 var
 	kind = require('../kind'),
-	utils = require('../utils'),
 	platform = require('../platform');
 
 var
