+++ lib/SpriteAnimation/SpriteAnimation.js
@@ -6,12 +6,12 @@ require('enyo');
 */
 
 var
-	kind = require('../kind');
+	kind = require('../kind'),
 	scope = window || global;
 
 var
 	Control = require('../Control'),
-	Image = require('../Image'),
+	EnyoImage = require('../Image'),
 	StylesheetSupport = require('../StylesheetSupport');
 
 /**
@@ -213,7 +213,7 @@ module.exports = kind(
 	* @private
 	*/
 	components: [
-		{kind: Image, name: 'spriteImage', classes: 'enyo-sprite-animation-image', mixins: [StylesheetSupport], sizing: 'cover'}
+		{kind: EnyoImage, name: 'spriteImage', classes: 'enyo-sprite-animation-image', mixins: [StylesheetSupport], sizing: 'cover'}
 	],
 
 	/**
