+++ lib/gesture/drag.js
@@ -33,7 +33,7 @@ var
 * @module enyo/gesture/drag
 * @public
 */
-var drag = module.exports = {
+module.exports = {
 
 	/**
 	* @private
