+++ lib/gesture/touchGestures.js
@@ -1,6 +1,5 @@
 var
 	dispatcher = require('../dispatcher'),
-	platform = require('../platform'),
 	utils = require('../utils');
 
 /**
@@ -18,7 +17,7 @@ var
 * @module enyo/gesture/touchGestures
 * @private
 */
-var touchGestures = module.exports = {
+module.exports = {
 
 	/**
 	* @private
